requirements.txt
fastapi[standard]
uvicorn
httpx
sqlmodel
python-dotenv
requests
pytest
pbcore-develop.zip